move two generic option creators into pybullet robot (#738)
