+++ tests/envs/test_pybullet_blocks.py
@@ -5,7 +5,7 @@ import pytest
 
 from predicators.src import utils
 from predicators.src.envs.pybullet_blocks import PyBulletBlocksEnv, \
-    _PyBulletState
+    PyBulletState
 from predicators.src.settings import CFG
 from predicators.src.structs import Object, State
 from predicators.tests.conftest import longrun
@@ -52,8 +52,7 @@ class _ExposedPyBulletBlocksEnv(PyBulletBlocksEnv):
         kinematics here.
         """
         joint_state = list(self._pybullet_robot.initial_joint_values)
-        state_with_sim = _PyBulletState(state.data,
-                                        simulator_state=joint_state)
+        state_with_sim = PyBulletState(state.data, simulator_state=joint_state)
         self._current_state = state_with_sim
         self._current_task = None
         self._reset_state(state_with_sim)
